+++ packages/scope-manager/src/lib/base-config.ts
@@ -1,7 +1,7 @@
 // THIS CODE WAS AUTOMATICALLY GENERATED
 // DO NOT EDIT THIS CODE BY HAND
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
-// npx nx generate-lib @typescript-eslint/scope-manager
+// npx nx generate-lib @typescript-eslint/repo-tools
 
 export const TYPE = Object.freeze({
   eslintImplicitGlobalSetting: 'readonly',