+++ packages/scope-manager/src/lib/decorators.ts
@@ -1,7 +1,7 @@
 // THIS CODE WAS AUTOMATICALLY GENERATED
 // DO NOT EDIT THIS CODE BY HAND
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
-// npx nx generate-lib @typescript-eslint/scope-manager
+// npx nx generate-lib @typescript-eslint/repo-tools
 
 import type { ImplicitLibVariableOptions } from '../variable';
 import { TYPE } from './base-config';
@@ -9,6 +9,8 @@ import { TYPE } from './base-config';
 export const decorators = {
   ClassMemberDecoratorContext: TYPE,
   DecoratorContext: TYPE,
+  DecoratorMetadataObject: TYPE,
+  DecoratorMetadata: TYPE,
   ClassDecoratorContext: TYPE,
   ClassMethodDecoratorContext: TYPE,
   ClassGetterDecoratorContext: TYPE,