+++ packages/scope-manager/src/lib/es2017.ts
@@ -1,10 +1,11 @@
 // THIS CODE WAS AUTOMATICALLY GENERATED
 // DO NOT EDIT THIS CODE BY HAND
 // RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
-// npx nx generate-lib @typescript-eslint/scope-manager
+// npx nx generate-lib @typescript-eslint/repo-tools
 
 import type { ImplicitLibVariableOptions } from '../variable';
 import { es2016 } from './es2016';
+import { es2017_date } from './es2017.date';
 import { es2017_intl } from './es2017.intl';
 import { es2017_object } from './es2017.object';
 import { es2017_sharedmemory } from './es2017.sharedmemory';
@@ -18,4 +19,5 @@ export const es2017 = {
   ...es2017_string,
   ...es2017_intl,
   ...es2017_typedarrays,
+  ...es2017_date,
 } as Record<string, ImplicitLibVariableOptions>;